refactor: v4.0.0/Switched type of averageAnswerTime to double also at serializer

+++ Server/Trivia_Server/JsonResponsePacketSerializer.h
@@ -130,7 +130,7 @@ typedef struct PlayerResults
     std::string username;
     unsigned int correctAnswerCount;
     unsigned int wrongAnswerCount;
-    unsigned int averageAnswerTime;
+    double averageAnswerTime;
 
 } PlayerResults;
